requirements.txt
# Django & DRF
Django>=5.0.1
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.1
drf-yasg>=1.21.7

# Database
psycopg2-binary>=2.9.9

# CUID Generation
cuid>=0.3

# Environment variables
python-dotenv>=1.0.0

# CORS
django-cors-headers>=4.3.1

# Production server
gunicorn>=21.2.0

# Development tools
black>=23.12.1
isort>=5.13.2
flake8>=7.0.0

# Image Processing
Pillow>=10.1.0 